The "Stephen obviously doesn't like the gem to build either" release.
Specifically, GitHub says "The gemspec is eval'd with $SAFE set to 3, disallowing usage of code like external requires and Dir.glob()." Using GemSpecOptions to build on GitHub doesn't work. I've created a rake task that builds a new gem spec based on the GemSpecOptions. This was it's still unified but static, the way GitHub requires it for the gem to actually build.